fix(common): tui editor ssr로 인한 빌드에러를 dynamic을 통해 고침
